==> code/getdata/1d_SPEI_from_GEE.ipynb <==
## SPEI data from google earth engine
import os
import ee
import geemap
import pandas as pd
import datetime
from datetime import date
import numpy as np

gridmet = ee.ImageCollection('GRIDMET/DROUGHT')

full_df = pd.read_csv("samples_to_get.csv")[["dnaSampleID", "collectDate", "long", "lat"]]

full_df

def get_spei(latitude_in, longitude_in, date_in, offset = 5, scale = 4000):
    end_date = datetime.date.fromisoformat(date_in)
    begin_date = end_date - datetime.timedelta(days=offset)   
    try:
        coord = ee.Geometry.Point(longitude_in, latitude_in)
        lst = gridmet.select('spi30d', 'spi90d', 'spi180d', 'spi270d',
         'spi1y', 'spi2y', 'spi5y', 'spei30d', 'spei90d', 'spei180d', 'spei270d',
                             'spei1y', 'spei2y', 'spei5y').filterDate( begin_date.isoformat(), end_date.isoformat())
        sample_spi_30d = lst.mean().sample(coord, scale).first().get('spi30d').getInfo()
        sample_spi_90d = lst.mean().sample(coord, scale).first().get('spi90d').getInfo()
        sample_spi_180d = lst.mean().sample(coord, scale).first().get('spi180d').getInfo()
        sample_spi_270d = lst.mean().sample(coord, scale).first().get('spi270d').getInfo()
        sample_spi_1 = lst.mean().sample(coord, scale).first().get('spi1y').getInfo()
        sample_spi_2 = lst.mean().sample(coord, scale).first().get('spi2y').getInfo()
        sample_spi_5 = lst.mean().sample(coord, scale).first().get('spi5y').getInfo()
        sample_spei_30d = lst.mean().sample(coord, scale).first().get('spei30d').getInfo()
        sample_spei_90d = lst.mean().sample(coord, scale).first().get('spei90d').getInfo()
        sample_spei_180d = lst.mean().sample(coord, scale).first().get('spei180d').getInfo()
        sample_spei_270d = lst.mean().sample(coord, scale).first().get('spei270d').getInfo()
        sample_spei_1 = lst.mean().sample(coord, scale).first().get('spei1y').getInfo()
        sample_spei_2 = lst.mean().sample(coord, scale).first().get('spei2y').getInfo()
        sample_spei_5 = lst.mean().sample(coord, scale).first().get('spei5y').getInfo()
        return(sample_spi_30d, sample_spi_90d, sample_spi_180d, 
               sample_spi_270d, sample_spi_1, sample_spi_2, sample_spi_5,
               sample_spei_30d, sample_spei_90d, sample_spei_180d, 
               sample_spei_270d, sample_spei_1, sample_spei_2, sample_spei_5)
    except:
        print("NA")
    

full_df[['spi30d', 'spi90d', 'spi180d', 'spi270d',
         'spi1y', 'spi2y', 'spi5y',
         'spei30d', 'spei90d', 'spei180d', 'spei270d',
                             'spei1y', 'spei2y', 'spei5y', 'test']] = np.nan

i = 0
for index, row in full_df.iterrows():
    latitude = row["lat"]
    longitude = row["long"]
    sampleDate = row["collectDate"]
    print(sampleDate)
    full_df.loc[i, ['spi30d', 'spi90d', 'spi180d', 'spi270d',
         'spi1y', 'spi2y', 'spi5y','spei30d', 'spei90d', 'spei180d', 'spei270d',
                    'spei1y', 'spei2y', 'spei5y']] = get_spei(latitude_in=latitude, 
                                                              longitude_in=longitude, 
                                                              date_in=sampleDate
                                                             )
    full_df.loc[i, 'test'] = longitude
    i += 1

full_df

full_df.to_csv("../../data/input/derived/spei.csv")
